fixed leaks and double frees. testcases are still the same. also tested the other mstest tests, with all of the tests we pass the most tests, looks good
